pyproject.toml
[project]
name = "fan-control"
version = "0.1.0"
description = "ML pipeline for thermal fan optimization"
readme = "README.md"
authors = [
    { name = "leonardcser", email = "73912641+leonardcser@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "scipy>=1.17.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "requests>=2.32.0",
]

[dependency-groups]
dev = [
    "torch>=2.9.1",
    "dvc>=3.66.1",
    "dvclive>=3.49.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "joblib>=1.4.0",
    "gpytorch>=1.15.1",
]

[project.scripts]
fan-control = "fan_control:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    "packages/data-collect",
]
